Move expr_to_str to its own sub-module
